+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added support for the Delete key, closes #297.
 - Added customizable support for the Backspace key.
 - Added support for Ctrl-A thru Z, closes #296.
+- Scrolling to the bottom of the terminal locks the scroll, closes #300.
+- Moving the scrollbar upwards in the Terminal pane breaks scroll lock, closes #267.
 
 ### Fixed